Fixed merge errors.
The function setPosition was already a defined as part of the Phaser3 Sprite class, so it was getting called with undefined values for gx and gy. I renamed it to setGamePosition. Also updated setTileContent to include all params in one { ... } notation so that they can be named when called for clarity.